telegram_bot/requirements.txt
aiogram
aiogram-dialog
dotenv
fastapi
httpx
pydantic-settings
pytz
PyYAML
redis
uvicorn